+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Changes this version:
 - Added basic translation function
 - Fixed CS selection in SPI buttons
+- Added axis output torque to FX live graph
 
 ### Changes in 1.14.x:
 - Added TMC space vector PWM checkbox